fix: handle position fixed mount
