+++ framework/changelog.mdx
@@ -5,9 +5,14 @@ route: /changelog
 
 # Changelog
 
+## 2.12.1
+
+- Fix for the case where `AMount` is `position: fixed`
+- Fix for the case where `AMount`/`AToastPlate` has toasts that overlap with an ascendant toast plate
+
 ## 2.12.0
 
-- Added, internally, `AToastPlate` to `AMount`.
+- Added, internally, `AToastPlate` to `AMount`
 
 ## 2.11.0
