+++ lib-demo/fixed-header.js
@@ -1,0 +1,112 @@
+import React, {useRef, useState} from "react";
+
+const App = () => {
+  const buttonRef = useRef(null);
+  const [menuOpen, setMenuOpen] = useState(false);
+  return (
+    <AApp>
+      <AMount style={{position: "fixed"}}>
+        <AHeader>
+          <AButton ref={buttonRef} onClick={() => setMenuOpen(!menuOpen)}>
+            Menu
+          </AButton>
+        </AHeader>
+        <AMenu anchorRef={buttonRef} open={menuOpen}>
+          <AListItem>List Item</AListItem>
+        </AMenu>
+      </AMount>
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+      <br />
+    </AApp>
+  );
+};
+
+export default App;
